fix(OverlayFS): return correct open flags in ReplyCreate
